🏗️ Architect: Modularized Conjunction Searches
Refactored the 691-line `apts/skyfield_searches/conjunctions.py` "God Object" into a modular package `apts/skyfield_searches/conjunctions/`.

Improvements:
- Decomposed monolithic search logic into specialized sub-modules:
    - `base.py`: Core search utilities and vectorized engines.
    - `planets.py`: Planet-to-planet and planet-to-sun conjunctions.
    - `lunar.py`: Lunar planetary occultations.
    - `objects.py`: Conjunctions with bright stars and Messier objects.
- Reduced cyclomatic complexity of `find_conjunctions_between_moving_bodies` by extracting helper functions.
- Maintained 100% backward compatibility via re-exports in `__init__.py`.
- Verified with full unit test suite (663 tests passed).
